LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/navigate-history-state-replace-expected.txt


FAIL history.state should be nulled by navigate() assert_equals: expected "#2" but got "#1"
